+++ src/contracts.ts
@@ -1246,6 +1246,16 @@ export interface ShellValueItem extends ValueItem {
 }
 
 /**
+ * Stores Slack API credentials.
+ */
+export interface SlackAPICredentials {
+    /**
+     * The access token.
+     */
+    token: string;
+}
+
+/**
  * A startup entry.
  */
 export type StartupEntry = string | StartupItem;
@@ -1722,6 +1732,11 @@ export const KEY_GLOBAL_SETTING_AZURE_DEVOPS_WORKSPACE_USERNAME = 'egoPTAzureDev
  * The key for the global setting that stores if CHANGELOG should be opened on startup automatically.
  */
 export const KEY_GLOBAL_SETTING_OPEN_CHANGELOG_ON_STARTUP = 'egoPTOpenChangelogOnStartup';
+
+/**
+ * The key for the global setting that stores a Slack API credentials.
+ */
+export const KEY_GLOBAL_SETTING_SLACK_API_CREDENTIALS = 'egoPTSlackAPICredentials';
 
 /**
  * The key of known TCP proxies.