+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log (vscode-powertools)
 
+## 0.32.0 (March 7th, 2019; tools)
+
+* can send files to [Slack](https://slack.com/) channels via `Power Tools: Tools` now
+* updated the following [npm](https://www.npmjs.com/) modules:
+  * [image-type](https://www.npmjs.com/package/image-type) `^3.1.0`
+
 ## 0.31.1 (March 5th, 2019; tools)
 
 * [$img](https://github.com/egodigital/vscode-powertools/wiki/Tools#imgdataorurl) function can handle URIs as input values now